Updating yumrepo configs for EPEL and Repoforge
RPMForge is now called Repoforge; the renaming has not progressed
through all the URLs yet.

Repoforge and EPEL also offer additional repositories that are not
enabled by default (e.g. "testing" and "extras").

I have parameterized this class so that the admin can toggle different
repositories on and off as desired.